alaa

+++ lib/Users/BalanceModels/get_card.dart
@@ -68,8 +68,8 @@ class User {
   City? gender;
   City? accountStatus;
   String? type;
-  int? availableBalance;
-  int? outstandingBalance;
+  double? availableBalance;
+  double? outstandingBalance;
 
   User(
       {this.id,
@@ -101,8 +101,8 @@ class User {
         ? new City.fromJson(json['account_status'])
         : null;
     type = json['type'];
-    availableBalance = json['available_balance'];
-    outstandingBalance = json['outstanding_balance'];
+    availableBalance = json['available_balance'].toDouble();
+    outstandingBalance = json['outstanding_balance'].toDouble();
   }
 
   Map<String, dynamic> toJson() {
